requirements.txt
antlr4-python3-runtime==4.9.2
certifi==2025.4.26
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
dotenv==0.9.9
packaging==25.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
